DeepMerge/pyproject.toml
[tool.poetry]
name = "cmlre-marine-data-platform"
version = "0.1.0"
description = "AI-Driven Unified Data Platform for Oceanography, Fisheries, and Molecular Biodiversity Insights (CMLRE – India)"
authors = ["CMLRE Team <team@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.2"
httpx = "^0.27.0"
anyio = "^4.4.0"
asyncpg = "^0.29.0"
sqlalchemy = {version = "^2.0.32", extras=["asyncio"]}
alembic = "^1.13.2"
psycopg-binary = "^3.2.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {version = "^1.7.4", extras=["bcrypt"]}
redis = "^5.0.7"
celery = "^5.4.0"
minio = "^7.2.7"
python-multipart = "^0.0.9"
limits = "^3.13.0"
slowapi = "^0.1.9"
starlette-graphene3 = "^0.6.0"
graphene = "^3.3"
prometheus-client = "^0.20.0"
loguru = "^0.7.2"
requests = "^2.32.5"
numpy = "^2.3.3"
netcdf4 = "^1.7.2"
xarray = "^2025.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"
ruff = "^0.6.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B"]
ignore = ["E203","E501"]
